final-pjt-back/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.2
cryptography==40.0.2
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.18
django-allauth==0.52.0
django-cors-headers==4.0.0
django-extensions==3.1.5
django-seed==0.3.1
djangorestframework==3.13.1
entrypoints==0.4
executing==1.2.0
Faker==13.3.4
fastjsonschema==2.16.2
fqdn==1.5.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.20.1
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
oauthlib==3.2.2
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2023.3
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
toposort==1.7
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2021.5
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
zipp==3.11.0
